int to long

+++ Project/chapter_05/src/show_earlierday.c
@@ -13,24 +13,24 @@
 
 int main(void)
   {
-    int d1, m1, y1;
-    int d2, m2, y2;
+    long d1, m1, y1;
+    long d2, m2, y2;
     long date1, date2;
 
     printf("Enter first date (mm/dd/yy): ");
-    scanf("%d/%d/%d", &m1, &d1, &y1);
+    scanf("%ld/%ld/%ld", &m1, &d1, &y1);
     printf("Enter second date (mm/dd/yy): ");
-    scanf("%d/%d/%d", &m2, &d2, &y2);
+    scanf("%ld/%ld/%ld", &m2, &d2, &y2);
 
     date1 = (long)(y1 * 10000 + m1 * 100 + d1);
     date2 = (long)(y2 * 10000 + m2 * 100 + d2);
 
     if (date1 < date2)
-      printf("%d/%d/%d is earlier than %d/%d/%d.\n", m1, d1, y1, m2, d2, y2);
+      printf("%ld/%ld/%ld is earlier than %ld/%ld/%ld.\n", m1, d1, y1, m2, d2, y2);
     else if (date1 > date2) 
-      printf("%d/%d/%d is earlier than %d/%d/%d.\n", m2, d2, y2, m1, d1, y1);
+      printf("%ld/%ld/%ld is earlier than %ld/%ld/%ld.\n", m2, d2, y2, m1, d1, y1);
     else
-      printf("%d/%d/%d is same as %d/%d/%d.\n", m2, d2, y2, m1, d1, y1);
+      printf("%ld/%ld/%ld is same as %ld/%ld/%ld.\n", m2, d2, y2, m1, d1, y1);
 
     return 0;
   }
